/usr/bin/time --format="%Es  %Mk" g++-5

```cpp
#include <brigand/sequences/back.hpp>
#include <brigand/sequences/list.hpp>
#include <brigand/sequences/make_sequence.hpp>
#include <brigand/types/integer.hpp>

template<class T> using void_t = T;

template<int N>
constexpr bool test_last_element()
{
  using list = brigand::push_back<brigand::make_sequence<brigand::size_t<0>, N>, int>;
  return brigand::back<list>(1);
}

static_assert(test_last_element<1>(), "");
static_assert(test_last_element<10>(), "");
static_assert(test_last_element<100>(), "");
static_assert(test_last_element<500>(), "");
static_assert(test_last_element<600>(), "");
static_assert(test_last_element<700>(), "");
static_assert(test_last_element<800>(), "");
static_assert(test_last_element<850>(), "");
//old: 0:00.80s  196232k
//new: 0:00.64s  156096k

static_assert(test_last_element<900>(), "");
static_assert(test_last_element<1000>(), "");
static_assert(test_last_element<2000>(), "");
static_assert(test_last_element<3000>(), "");
//old: 0:05.70s  1009680k
//new: 0:03.85s  774852k
```