@edouarda 2016-02-25
MSVC 2013 doesn't like it! :-)