@porkybrain 2016-02-25
MSVC2013 has problems with templated aliases sometimes. 

This should work though:

```C++
		template<class> 
		struct voidp{
			using type = void const *;
		};

		template<class T, class... Ts>
		struct last_element
		{
			template<class Us>
			static Us last(typename voidp<Ts>::type..., Us*);

			using type = decltype(last(nullptr, static_cast<Ts*>(nullptr)...));
		};
```