@porkybrain 2016-02-25
actually since we already have an optimized `at` why not just use that?