@jonathanpoelen 2016-02-25
I did not think about it :)

After trying with `brigand::at_c<list, N>`, I get:
- 0:01.69s  329400k
- 0:09.14s  1656948k.

Much slower than a specialized version. I will work on improving `at_c` before resuming `last_element`.