@jfalcou 2016-02-25
last_element could be sooo simple if decltype(declval<T>()...)  was allowed :(